dashboard/requirements.txt
streamlit
psycopg2
python-dotenv
pandas
matplotlib
seaborn
boto3